+++ l10n/nl.js
@@ -124,5 +124,7 @@ OC.L10N.register("linkboard", {
     "Invalid JSON in payload: %s": "Ongeldige JSON: %s",
     "Invalid JSON": "Ongeldige JSON",
     "Import failed: %s": "Import mislukt: %s",
-    "Maximum number of services reached (%s)": "Maximaal aantal services bereikt (%s)"
+    "Maximum number of services reached (%s)": "Maximaal aantal services bereikt (%s)",
+    "Display": "Display",
+    "Show scrollbar": "Show scrollbar"
 }, "nplurals=2; plural=(n != 1);");