+++ l10n/zh_HK.js
@@ -124,5 +124,7 @@ OC.L10N.register("linkboard", {
     "Invalid JSON in payload: %s": "无效的 JSON：%s",
     "Invalid JSON": "无效的 JSON",
     "Import failed: %s": "导入失败：%s",
-    "Maximum number of services reached (%s)": "已达到服务最大数量（%s）"
+    "Maximum number of services reached (%s)": "已达到服务最大数量（%s）",
+    "Display": "Display",
+    "Show scrollbar": "Show scrollbar"
 }, "nplurals=1; plural=0;");