+++ src/components/Dashboard/DashboardView.vue
@@ -653,6 +653,7 @@ export default {
         background-size: cover;
         background-position: center;
         background-repeat: no-repeat;
+        pointer-events: none;
     }
 
     &__content {